requirements.txt
astroid==2.4.2
autopep8==1.5.4
colorama==0.4.4
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysql-connector-python==8.0.22
prettytable==1.0.1
protobuf==3.13.0
pycodestyle==2.6.0
pylint==2.6.0
six==1.15.0
toml==0.10.1
wcwidth==0.2.5
wrapt==1.12.1
